+++ src/components/Audioplayer/TokenSegments/style.scss
@@ -1,0 +1,2 @@
+@import '../../../styles/_bootstrap-config.scss';
+
