+++ src/components/Audioplayer/Track/index.js
@@ -8,7 +8,6 @@ import Tracker from './Tracker';
 import { clearCurrentWord, setCurrentWord } from '../../../redux/modules/ayahs';
 
 const style = require('./style.scss');
-const segments = require('./segments.js');
 
 @connect(
   (state, ownProps) => {
@@ -40,43 +39,9 @@ export default class Track extends Component {
     listeners: {}
   };
 
-  constructor() {
-    super(...arguments);
-    this.cues = [];
-    this.seek = {}
-    segments.forEach((row, id) => {
-      const cols = row.split(/\t/);
-      const surahId = cols[0];
-      const ayahNum = cols[1];
-      const start = parseInt(cols[2], 10);
-      const duration = parseInt(cols[3], 10);
-      const tokenId = parseInt(cols[4], 10);
-      if (typeof this.cues[surahId] == 'undefined')
-        this.cues[surahId] = [];
-      if (typeof this.cues[surahId][ayahNum] == 'undefined')
-        this.cues[surahId][ayahNum] = [];
-      const data = { start, duration, tokenId };
-      this.cues[surahId][ayahNum].push(data);
-      const key = surahId +':'+ ayahNum +':'+ tokenId;
-      this.seek[key] = this.seek[key]? this.seek[key] : { ...data, cueSequence: this.cues[surahId][ayahNum].length - 1 };
-    });
-    //console.log('constructor', this.cues);
-  }
-
-  resetHighlightProps() {
-    //console.log('resetHighlightProps');
-    this.singleMode = false;
-    this.totalHighlightedDuration = 0;
-    clearTimeout(this.timerId1);
-    clearTimeout(this.timerId2);
-    clearTimeout(this.timerId3);
-    //this.clearCurrentWord();
-  }
-
   componentDidMount() {
     if (this.props.file) {
       this.onFileLoad(this.props.file);
-      this.resetHighlightProps();
     }
   }
 
@@ -101,48 +66,6 @@ export default class Track extends Component {
   }
 
   componentWillReceiveProps(nextProps) {
-    let getSegment = () => {
-      return (
-       this.props.currentWord !== nextProps.currentWord
-       && (nextProps.currentWord !== this.state.currentWord) // || this.props.currentWord !== this.state.currentWord)
-       && this.seek[nextProps.currentWord]
-    )};
-
-    let seekToSegment = (file, currentWord, segment) => {
-      const singleMode = this.singleMode;
-      file.pause();
-      this.singleMode = singleMode; // because pause resets singleMode?
-      const events = ['loadeddata', 'loaded', 'load', 'canplay', 'canplaythrough', 'loadstart'];
-      let setTime;
-      setTime = (now = false) => {
-        if (now === true) {
-          this.totalHighlightedDuration = segment.start;
-          file.currentTime = segment.start / 1000.0;
-          //console.log('setTime immediate', { currentWord, src: file.src, currentTime: file.currentTime, singleMode: this.singleMode, isPlaying: this.props.isPlaying });
-          this.setState({ currentWord });
-        } else
-        if (currentWord === this.props.currentWord) {
-          this.totalHighlightedDuration = segment.start;
-          file.currentTime = segment.start / 1000.0;
-          //console.log('setTime not immediate', { currentWord, src: file.src, currentTime: file.currentTime, singleMode: this.singleMode, isPlaying: this.props.isPlaying });
-          this.setState({ currentWord });
-        }
-        events.every((evName) => {
-          file.removeEventListener(evName, setTime);
-          return true;
-        });
-      };
-      if (file.readyState >= 4) {
-        setTime(true);
-      } else {
-        events.every((evName) => {
-          file.addEventListener(evName, setTime, false);
-          return true;
-        });
-      }
-     //file.play(); // ???
-    };
-
     if (!nextProps.file) return;
     if (this.props.file.src !== nextProps.file.src) {
       this.onFileUnload(this.props.file);
@@ -151,13 +74,6 @@ export default class Track extends Component {
       });
 
       this.onFileLoad(nextProps.file);
-      this.resetHighlightProps();
-    }
-
-    const segment = getSegment();
-    if (segment) {
-      //console.log('seekToSegment', { propsWord: this.props.currentWord, stateWord: this.state.currentWord, file: nextProps.file, nextWord: nextProps.currentWord, segment });
-      seekToSegment(nextProps.file, nextProps.currentWord, segment);
     }
   }
 
@@ -218,123 +134,25 @@ export default class Track extends Component {
         currentTime
       });
 
-      this.nourPlay();
-
       onPlay();
     };
     file.addEventListener('play', play, false);
-
-    const pause = () => {
-      const { onPause } = this.props;
-
-      //console.log('file.addEventListener(pause)', { file, state: this.state });
-
-      this.nourPause();
-
-      //onPause();
-    };
-    file.addEventListener('pause', pause, false);
 
     this.setState({
       listeners: {
         loadeddata,
         timeupdate,
         ended,
-        play,
-        pause
+        play
       }
     });
   }
 
   onFileUnload(file) {
     this.props.file.pause();
-    [ 'loadeddata', 'timeupdate', 'ended', 'play', 'pause' ].forEach((listener) => {
+    [ 'loadeddata', 'timeupdate', 'ended', 'play' ].forEach((listener) => {
       file.removeEventListener(listener, this.state.listeners[listener]);
     });
-  }
-
-  nourPlay() {
-    const { surahId, ayahNum, currentWord } = this.props;
-    let cueSequence = 0;
-    if (this.seek[currentWord]) {
-      this.totalHighlightedDuration = this.seek[currentWord].start;
-      cueSequence = this.seek[currentWord].cueSequence;
-    }
-    //console.log('nourPlay', { highlightedTime: this.totalHighlightedDuration, surahId, ayahNum, currentWord, cueSequence, stateWord: this.state.currentWord });
-
-    this.highlightToken(surahId, ayahNum, cueSequence, 0);
-  }
-
-  nourPause() {
-    const { surahId, ayahNum, currentWord } = this.props;
-    //console.log('nourPause', { surahId, ayahNum, currentWord });
-
-    this.resetHighlightProps();
-  }
-
-  setCurrentWord(surahId, ayahNum, tokenId) {
-    const currentWord = `${surahId}:${ayahNum}:${tokenId}`;
-    const { setCurrentWord } = this.props;
-    this.setState({ currentWord });
-    setCurrentWord(currentWord);
-  }
-
-  clearCurrentWord() {
-    const { clearCurrentWord } = this.props;
-    this.setState({ currentWord: null });
-    clearCurrentWord();
-  }
-
-  highlightToken(surahId, ayahNum, cueSequence, syncError) {
-    const seq = this.cues[surahId][ayahNum][cueSequence];
-    const { file, setCurrentWord } = this.props;
-
-    //console.log('highlightToken', { surahId, ayahNum, cueSequence, syncError, setCurrentWord, file, seq });
-
-    if (!seq) return;
-    if (seq.tokenId >= 0) {
-      //console.log('this.setCurrentWord', { surahId, ayahNum, tokenId: seq.tokenId, cueSequence, seq });
-      this.setCurrentWord(surahId, ayahNum, seq.tokenId);
-    } else {
-      // silence.show(); ???
-    }
-
-    this.timerId1 = setTimeout(() => {
-      this.totalHighlightedDuration += seq.duration;
-      syncError = this.totalHighlightedDuration - file.currentTime * 1000
-
-      //console.log("timerId1", { highlightedTime: this.totalHighlightedDuration, currentTime: file.currentTime * 1000, cueSequence, syncError })
-
-      if (syncError >= 100) {
-        this.timerId2 = setTimeout(() => {
-          //console.log("timerId2", { highlightedTime: this.totalHighlightedDuration, currentTime: file.currentTime * 1000, cueSequence, syncError })
-          this.unhighlightToken(surahId, ayahNum, cueSequence, 0);
-        }, syncError);
-      } else {
-        this.unhighlightToken(surahId, ayahNum, cueSequence, syncError);
-      }
-    }, seq.duration + syncError);
-  }
-
-  unhighlightToken(surahId, ayahNum, cueSequence, syncError) {
-    const seq = this.cues[surahId][ayahNum][cueSequence];
-    const { file, currentWord } = this.props;
-
-    //console.log('unhighlightToken', { currentWord, cueSequence, syncError });
-
-    if (seq.tokenId >= 0) {
-      this.clearCurrentWord();
-    } else {
-      //silence.hide() // ???
-    }
-
-    if (!this.singleMode) {
-      this.highlightToken(surahId, ayahNum, cueSequence + 1, syncError);
-    } else {
-      this.file.pause();
-    }
-
-    this.singleMode = false;
   }
 
   onTrackerMove(event) {
